Wisteria/requirements.txt
pytesseract==0.3.8
Pillow==9.3.0
opencv-python==4.6.0.66
numpy==1.23.5
pandas==1.5.2
regex==2022.10.31
datetime==4.6